feat: add encode client and embedding cache to PD worker
