+++ components/src/dynamo/vllm/args.py
@@ -79,7 +79,6 @@ def parse_args() -> Config:
     Returns:
         Config: Parsed configuration object.
     """
-
     dynamo_runtime_argspec = DynamoRuntimeArgGroup()
     dynamo_vllm_argspec = DynamoVllmArgGroup()
