+++ components/src/dynamo/vllm/multimodal_handlers/__init__.py
@@ -5,14 +5,14 @@ from dynamo.vllm.multimodal_handlers.encode_worker_handler import (
     EncodeWorkerHandler,
     VLLMEncodeWorkerHandler,
 )
+from dynamo.vllm.multimodal_handlers.multimodal_pd_worker_handler import (
+    MultimodalPDWorkerHandler,
+)
 from dynamo.vllm.multimodal_handlers.preprocessed_handler import (
     ECProcessorHandler,
     PreprocessedHandler,
 )
-from dynamo.vllm.multimodal_handlers.worker_handler import (
-    MultimodalDecodeWorkerHandler,
-    MultimodalPDWorkerHandler,
-)
+from dynamo.vllm.multimodal_handlers.worker_handler import MultimodalDecodeWorkerHandler
 
 __all__ = [
     "EncodeWorkerHandler",
